content/xul/document/test/test_bug497875.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=497875
-->
<window title="Mozilla Bug 497875"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

<body  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=497875">Mozilla Bug 497875</a>
<p id="display"><iframe id="iframe" src="bug497875-iframe.xul"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
</pre>
</body>

<script class="testbody" type="application/javascript">
<![CDATA[

/** Test for Bug 497875 **/

SimpleTest.waitForExplicitFinish();

function done_test() {
  var f = document.getElementById('iframe')
  f.src="about:blank"
  f.parentNode.removeChild(f)
  ok(true, "this is a crashtest");
  SimpleTest.finish();
}

function do_test() {
  setTimeout(function () {document.getElementById('iframe').contentWindow.location.reload()}, 500);
  setTimeout(function () {document.getElementById('iframe').contentWindow.location.reload()}, 1000);
  setTimeout(function () {document.getElementById('iframe').contentWindow.location.reload()}, 1500);
  setTimeout(done_test, 2000);
}

do_test();


]]>
</script>

</window>
